requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
click==8.1.7
    # via -r requirements.in
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via -r requirements.in
ruff==0.4.6
    # via -r requirements.in
smmap==5.0.1
    # via gitdb
